Some fabric refactoring, add support for fabric 26.2
