+++ mod-fabric/src/main/resources/fabric.mod.json
@@ -30,7 +30,10 @@
       "1.21.8",
       "1.21.9",
       "1.21.10",
-      "1.21.11"
+      "1.21.11",
+      "26.1.1",
+      "26.1.2",
+      "26.2"
     ],
     "fabric-api": "*",
     "fabric-permissions-api-v0": "*",
