+++ platform-fabric-2612/src/main/java/de/pianoman911/playerculling/platformfabric2612/common/IServerPlayer.java
@@ -1,7 +1,7 @@
-package de.pianoman911.playerculling.platformfabric261.common;
+package de.pianoman911.playerculling.platformfabric2612.common;
 // Created by booky10 in PlayerCulling (18:15 14.07.2025)
 
-import de.pianoman911.playerculling.platformfabric261.platform.FabricPlayer;
+import de.pianoman911.playerculling.platformfabric2612.platform.FabricPlayer;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
