+++ platform-fabric-2612/src/main/java/de/pianoman911/playerculling/platformfabric2612/platform/FabricCommandSender.java
@@ -1,4 +1,4 @@
-package de.pianoman911.playerculling.platformfabric261.platform;
+package de.pianoman911.playerculling.platformfabric2612.platform;
 
 import de.pianoman911.playerculling.platformcommon.platform.IPlatform;
 import de.pianoman911.playerculling.platformcommon.platform.command.PlatformCommandSender;
