+++ platform-fabric-2612/src/main/java/de/pianoman911/playerculling/platformfabric2612/util/BlockStateUtil.java
@@ -1,4 +1,4 @@
-package de.pianoman911.playerculling.platformfabric261.util;
+package de.pianoman911.playerculling.platformfabric2612.util;
 
 import de.pianoman911.playerculling.platformcommon.util.OcclusionMappings;
 import net.minecraft.core.BlockPos;
