+++ platform-fabric-2612/src/main/resources/playerculling.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "de.pianoman911.playerculling.platformfabric12111.mixin",
+  "package": "de.pianoman911.playerculling.platformfabric2612.mixin",
   "compatibilityLevel": "JAVA_25",
   "mixins": [
     "ChunkMapMixin",
